requirements.txt
google-generativeai
requests
gtts
PyYAML
beautifulsoup4
readability-lxml
pillow
pytesseract
streamlit
openai
google-cloud-texttospeech
python-dotenv 